Participant class has been moved to its own file.
